pyproject.toml
[project]
name = "jpg-convertor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "typer>=0.16.1",
]
